requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
langchain-core==0.3.42
langchain-groq==0.2.5
tavily-python==0.3.1
pandas==2.0.2
pydantic==2.7.4
httpx==0.27.0
python-multipart==0.0.7